Move Col CRDT to core.crdt
